requirements.txt
# ── Core LLM ─────────────────────────────────────────────────────────────────
groq>=0.9.0

# ── Vector Store ─────────────────────────────────────────────────────────────
faiss-cpu>=1.8.0
numpy>=1.26.0

# ── Web API ───────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
websockets>=12.0
httpx>=0.27.0
python-multipart>=0.0.9

# ── Security & Crypto ────────────────────────────────────────────────────────
cryptography>=42.0.0

# ── Config ───────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0

# ── Logging ──────────────────────────────────────────────────────────────────
structlog>=24.1.0

# ── Terminal Dashboard ───────────────────────────────────────────────────────
rich>=13.7.0

# ── Testing ──────────────────────────────────────────────────────────────────
hypothesis>=6.100.0
pytest>=8.0.0
pytest-mock>=3.14.0
pytest-asyncio>=0.23.0

# ── Optional: PDF Export ─────────────────────────────────────────────────────
# pip install reportlab

# ── Optional: SMS Alerts ─────────────────────────────────────────────────────
# pip install twilio
